rag_service/requirements.txt
chromadb==0.5.5
google-genai>=1.2.0
langchain
langchain-google-genai
python-dotenv
flask
flask-cors
pypdf
